Authentication added

+++ cafe/grails-app/domain/billing/Order.groovy
@@ -6,7 +6,7 @@ import user.Employee
 class Order {
     static belongsTo = [employee: Employee]
     Date orderDate
-    int amount
+    double amount
     static constraints = {
     }
 }
